+++ vite.config.js
@@ -15,4 +15,14 @@ export default defineConfig({
     esbuild: {
         jsx: 'automatic',
     },
+    build: {
+        rollupOptions: {
+            output: {
+                manualChunks: {
+                    react: ['react', 'react-dom'],
+                    bootstrap: ['bootstrap'],
+                },
+            },
+        },
+    },
 });